requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
gcloud==0.18.3
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
jws==0.1.3
oauth2client==4.1.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pycryptodomex==3.18.0
pyparsing==3.0.9
Pyrebase4==4.7.1
python-jwt==2.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
